Python client cleanup: namespaces and test separtation.
