+++ datajunction-clients/python/datajunction/_internal.py
@@ -34,8 +34,6 @@ _logger = logging.getLogger(__name__)
 #
 # Helpers
 #
-
-
 def from_jupyter() -> bool:  # pragma: no cover
     """
     Checks whether we're running from an IPython interactive console
@@ -227,34 +225,6 @@ class DJClient:
                 f"A node with name {node_name} exists, but it is not a {type_} node!",
             )
         return node
-
-    def _delete_node(self, node_name: str) -> None:
-        """
-        Delete (aka deactivate) this node.
-        """
-        response = self._session.delete(
-            f"/nodes/{node_name}/",
-            timeout=self._timeout,
-        )
-        json_response = response.json()
-        if not response.ok:
-            raise DJClientException(
-                f"Deleting node `{node_name}` failed: {json_response}",
-            )  # pragma: no cover
-
-    def _restore_node(self, node_name: str) -> None:
-        """
-        Restore (aka reactivate) this node.
-        """
-        response = self._session.post(
-            f"/nodes/{node_name}/restore/",
-            timeout=self._timeout,
-        )
-        json_response = response.json()
-        if not response.ok:
-            raise DJClientException(
-                f"Restoring node `{node_name}` failed: {json_response}",
-            )  # pragma: no cover
 
     def _validate_node(self, node: "Node"):
         """
