+++ datajunction-clients/python/tests/test_integration.py
@@ -136,6 +136,7 @@ def test_integration():  # pylint: disable=too-many-statements,too-many-locals,l
             "default.avg_repair_price",
             "default.total_repair_cost",
         ],
+        name_only=True,
     )
     assert common_dimensions == [
         "default.dispatcher.company_name",
@@ -169,7 +170,7 @@ def test_integration():  # pylint: disable=too-many-statements,too-many-locals,l
         "default.us_state.state_region_description",
     ]
     # with details
-    common_dimensions = dj.common_dimensions_with_details(
+    common_dimensions = dj.common_dimensions(
         metrics=[
             "default.num_repair_orders",
             "default.avg_repair_price",
@@ -413,6 +414,7 @@ def test_integration():  # pylint: disable=too-many-statements,too-many-locals,l
     # names only
     common_metrics = dj.common_metrics(
         dimensions=["default.date_dim", "default.repair_order"],
+        name_only=True,
     )
     assert common_metrics == [
         "default.num_repair_orders",
@@ -423,213 +425,45 @@ def test_integration():  # pylint: disable=too-many-statements,too-many-locals,l
         "default.avg_time_to_dispatch",
     ]
     # with details
-    common_metrics = dj.common_metrics_with_details(
+    common_metrics = dj.common_metrics(
         dimensions=["default.date_dim", "default.repair_order"],
     )
     assert common_metrics == [
         {
-            "node_revision_id": 23,
-            "node_id": 23,
-            "type": "metric",
             "name": "default.num_repair_orders",
             "display_name": "Default: Num Repair Orders",
-            "version": "v1.0",
-            "status": "valid",
-            "mode": "published",
-            "catalog": {
-                "id": 1,
-                "uuid": "48474405-6e39-46f8-88b1-0bea868aa947",
-                "created_at": "2023-07-22T00:02:43.554751+00:00",
-                "updated_at": "2023-07-22T00:02:43.554752+00:00",
-                "extra_params": {},
-                "name": "warehouse",
-            },
-            "schema_": None,
-            "table": None,
             "description": "Number of repair orders",
             "query": "SELECT  count(repair_order_id) default_DOT_num_repair_orders \n FROM default.repair_orders\n\n",
-            "availability": None,
-            "columns": [
-                {
-                    "name": "default_DOT_num_repair_orders",
-                    "type": "bigint",
-                    "attributes": [],
-                    "dimension": None,
-                },
-            ],
-            "updated_at": "2023-07-22T00:02:51.460733+00:00",
-            "materializations": [],
-            "parents": [{"name": "default.repair_orders"}],
         },
         {
-            "node_revision_id": 24,
-            "node_id": 24,
-            "type": "metric",
             "name": "default.avg_repair_price",
             "display_name": "Default: Avg Repair Price",
-            "version": "v1.0",
-            "status": "valid",
-            "mode": "published",
-            "catalog": {
-                "id": 1,
-                "uuid": "48474405-6e39-46f8-88b1-0bea868aa947",
-                "created_at": "2023-07-22T00:02:43.554751+00:00",
-                "updated_at": "2023-07-22T00:02:43.554752+00:00",
-                "extra_params": {},
-                "name": "warehouse",
-            },
-            "schema_": None,
-            "table": None,
             "description": "Average repair price",
             "query": "SELECT  avg(price) default_DOT_avg_repair_price \n FROM default.repair_order_details\n\n",
-            "availability": None,
-            "columns": [
-                {
-                    "name": "default_DOT_avg_repair_price",
-                    "type": "double",
-                    "attributes": [],
-                    "dimension": None,
-                },
-            ],
-            "updated_at": "2023-07-22T00:02:51.616490+00:00",
-            "materializations": [],
-            "parents": [{"name": "default.repair_order_details"}],
         },
         {
-            "node_revision_id": 25,
-            "node_id": 25,
-            "type": "metric",
             "name": "default.total_repair_cost",
             "display_name": "Default: Total Repair Cost",
-            "version": "v1.0",
-            "status": "valid",
-            "mode": "published",
-            "catalog": {
-                "id": 1,
-                "uuid": "48474405-6e39-46f8-88b1-0bea868aa947",
-                "created_at": "2023-07-22T00:02:43.554751+00:00",
-                "updated_at": "2023-07-22T00:02:43.554752+00:00",
-                "extra_params": {},
-                "name": "warehouse",
-            },
-            "schema_": None,
-            "table": None,
             "description": "Total repair cost",
             "query": "SELECT  sum(price) default_DOT_total_repair_cost \n FROM default.repair_order_details\n\n",
-            "availability": None,
-            "columns": [
-                {
-                    "name": "default_DOT_total_repair_cost",
-                    "type": "double",
-                    "attributes": [],
-                    "dimension": None,
-                },
-            ],
-            "updated_at": "2023-07-22T00:02:51.834912+00:00",
-            "materializations": [],
-            "parents": [{"name": "default.repair_order_details"}],
         },
         {
-            "node_revision_id": 27,
-            "node_id": 27,
-            "type": "metric",
             "name": "default.total_repair_order_discounts",
             "display_name": "Default: Total Repair Order Discounts",
-            "version": "v1.0",
-            "status": "valid",
-            "mode": "published",
-            "catalog": {
-                "id": 1,
-                "uuid": "48474405-6e39-46f8-88b1-0bea868aa947",
-                "created_at": "2023-07-22T00:02:43.554751+00:00",
-                "updated_at": "2023-07-22T00:02:43.554752+00:00",
-                "extra_params": {},
-                "name": "warehouse",
-            },
-            "schema_": None,
-            "table": None,
             "description": "Total repair order discounts",
             "query": "SELECT  sum(price * discount) default_DOT_total_repair_order_discounts \n FROM default.repair_order_details\n\n",
-            "availability": None,
-            "columns": [
-                {
-                    "name": "default_DOT_total_repair_order_discounts",
-                    "type": "double",
-                    "attributes": [],
-                    "dimension": None,
-                },
-            ],
-            "updated_at": "2023-07-22T00:02:52.276818+00:00",
-            "materializations": [],
-            "parents": [{"name": "default.repair_order_details"}],
         },
         {
-            "node_revision_id": 28,
-            "node_id": 28,
-            "type": "metric",
             "name": "default.avg_repair_order_discounts",
             "display_name": "Default: Avg Repair Order Discounts",
-            "version": "v1.0",
-            "status": "valid",
-            "mode": "published",
-            "catalog": {
-                "id": 1,
-                "uuid": "48474405-6e39-46f8-88b1-0bea868aa947",
-                "created_at": "2023-07-22T00:02:43.554751+00:00",
-                "updated_at": "2023-07-22T00:02:43.554752+00:00",
-                "extra_params": {},
-                "name": "warehouse",
-            },
-            "schema_": None,
-            "table": None,
             "description": "Total repair order discounts",
             "query": "SELECT  avg(price * discount) default_DOT_avg_repair_order_discounts \n FROM default.repair_order_details\n\n",
-            "availability": None,
-            "columns": [
-                {
-                    "name": "default_DOT_avg_repair_order_discounts",
-                    "type": "double",
-                    "attributes": [],
-                    "dimension": None,
-                },
-            ],
-            "updated_at": "2023-07-22T00:02:52.428709+00:00",
-            "materializations": [],
-            "parents": [{"name": "default.repair_order_details"}],
         },
         {
-            "node_revision_id": 29,
-            "node_id": 29,
-            "type": "metric",
             "name": "default.avg_time_to_dispatch",
             "display_name": "Default: Avg Time To Dispatch",
-            "version": "v1.0",
-            "status": "valid",
-            "mode": "published",
-            "catalog": {
-                "id": 1,
-                "uuid": "48474405-6e39-46f8-88b1-0bea868aa947",
-                "created_at": "2023-07-22T00:02:43.554751+00:00",
-                "updated_at": "2023-07-22T00:02:43.554752+00:00",
-                "extra_params": {},
-                "name": "warehouse",
-            },
-            "schema_": None,
-            "table": None,
             "description": "Average time to dispatch a repair order",
             "query": "SELECT  avg(dispatched_date - order_date) default_DOT_avg_time_to_dispatch \n FROM default.repair_orders\n\n",
-            "availability": None,
-            "columns": [
-                {
-                    "name": "default_DOT_avg_time_to_dispatch",
-                    "type": "date",
-                    "attributes": [],
-                    "dimension": None,
-                },
-            ],
-            "updated_at": "2023-07-22T00:02:52.617142+00:00",
-            "materializations": [],
-            "parents": [{"name": "default.repair_orders"}],
         },
     ]
